backend/requirements.txt
fastapi==0.111.0
motor==3.4.0
python-decouple==3.8
uvicorn==0.30.1
